Made requested changed to twap and removed starting delay
