+++ hummingbot/strategy/dev_4_twap/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+from .dev_4_twap import Dev4TwapTradeStrategy
+
+
+__all__ = [
+    Dev4TwapTradeStrategy
+]
