+++ hummingbot/strategy/dev_4_twap/dev_4_twap_config_map.py
@@ -10,7 +10,7 @@ from hummingbot.client.settings import (
 
 
 def symbol_prompt():
-    market = execution4_config_map.get("market").value
+    market = dev_4_twap_config_map.get("market").value
     example = EXAMPLE_PAIRS.get(market)
     return "Enter the token symbol you would like to trade on %s%s >>> " \
            % (market, f" (e.g. {example})" if example else "")
@@ -22,11 +22,11 @@ def str2bool(value: str):
 
 # checks if the symbol pair is valid
 def is_valid_market_trading_pair_tuple(value: str) -> bool:
-    market = execution4_config_map.get("market").value
+    market = dev_4_twap_config_map.get("market").value
     return is_valid_market_symbol(market, value)
 
 
-execution4_config_map = {
+dev_4_twap_config_map = {
     "market":
         ConfigVar(key="market",
                   prompt="Enter the name of the exchange >>> ",
@@ -53,28 +53,28 @@ execution4_config_map = {
                   prompt="Enter True for Buy order and False for Sell order (default is Buy Order) >>> ",
                   type_str="bool",
                   default=True),
-    "time_delay":
-        ConfigVar(key="time_delay",
-                  prompt="How many seconds do you want to wait between each individual order? (Enter 10 to indicate 10 seconds. "
-                         "Default is 10)? >>> ",
-                  type_str="float",
-                  default=10),
     "num_individual_orders":
         ConfigVar(key="num_individual_orders",
                   prompt="Into how many individual orders do you want to split this order? (Enter 10 to indicate 10 individual orders. "
                          "Default is 1)? >>> ",
                   type_str="float",
                   default=1),
+    "time_delay":
+        ConfigVar(key="time_delay",
+                  prompt="How many seconds do you want to wait between each individual order? (Enter 10 to indicate 10 seconds. "
+                         "Default is 10)? >>> ",
+                  type_str="float",
+                  default=10),
     "order_price":
         ConfigVar(key="order_price",
                   prompt="What is the price of the limit order ? >>> ",
-                  required_if=lambda: execution4_config_map.get("order_type").value == "limit",
+                  required_if=lambda: dev_4_twap_config_map.get("order_type").value == "limit",
                   type_str="float"),
     "cancel_order_wait_time":
         ConfigVar(key="cancel_order_wait_time",
                   prompt="How long do you want to wait before cancelling your limit order (in seconds). "
                          "(Default is 60 seconds) ? >>> ",
-                  required_if=lambda: execution4_config_map.get("order_type").value == "limit",
+                  required_if=lambda: dev_4_twap_config_map.get("order_type").value == "limit",
                   type_str="float",
                   default=60),
