--- hummingbot/strategy/execution4/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-
-from .execution4 import Execution4TradeStrategy
-
-
-__all__ = [
-    Execution4TradeStrategy
-]
