+++ hummingbot/templates/conf_dev_4_twap_strategy_TEMPLATE.yml
@@ -3,7 +3,7 @@
 ########################################################
 
 # The following configuations are only required for the
-# simple trade strategy
+# twap strategy
 
 # Exchange and token parameters
 market: null
@@ -22,11 +22,11 @@ order_price: null
 # it will cancel the limit order.
 cancel_order_wait_time: null
 
-# How long to between placing incremental orders
-time_delay: null
-
 # How many orders to place
 num_individual_orders: null
+
+# How long to between placing incremental orders
+time_delay: null
 
 # Specify buy/sell. Set to True for Buy and False for Sell
 is_buy: null
